requirements.txt
pyqt5==5.15.2
python-dotenv==0.19.2